+++ src/smw/world/blocks/FlipBlock.java
@@ -27,7 +27,7 @@ public  class FlipBlock extends AnimatedBlock{
   }
   
   @Override
-  public int collideWithLeft(Player player, int newX) {
+  public float collideWithLeft(Player player, float newX) {
     if(running){
       return newX;
     }
@@ -36,7 +36,7 @@ public  class FlipBlock extends AnimatedBlock{
   }
 
   @Override
-  public int collideWithRight(Player player, int newX) {
+  public float collideWithRight(Player player, float newX) {
     if(running){
       return newX;
     }
@@ -45,7 +45,7 @@ public  class FlipBlock extends AnimatedBlock{
   }
 
   @Override
-  public int collideWithTop(Player player, int newY) {
+  public float collideWithTop(Player player, float newY) {
     if(running){
       return newY;
     }
@@ -54,7 +54,7 @@ public  class FlipBlock extends AnimatedBlock{
   }
 
   @Override
-  public int collideWithBottom(Player player, int newY) {    
+  public float collideWithBottom(Player player, float newY) {    
     if(running){
       return newY;
     }
